refactor(LegIcon): Rename LegIconCore to LegIcon, delete the other one.
